chore(deps): swap fuzzywuzzy for rapidfuzz in app code (#170)
All nine fuzzywuzzy sites use only plain `fuzz.ratio(a, b)` -- identical
signature in rapidfuzz -- so the swap is mechanical:

- seven files with live usage move to `from rapidfuzz import fuzz`
  (ActionChooser also drops its unused `process` import);
- Store.py and StorePage.py had dead imports, deleted rather than converted;
- PageSelector.calc_ratio's return annotation becomes `-> float`, since
  rapidfuzz returns a float in [0, 100] where fuzzywuzzy rounded to int. The
  `@staticmethod` + `@lru_cache` shape is deliberate and stays as-is.

Every consumer was audited for the int->float change: threshold filters
compare fine, the GTK sort funcs return -1/0/1 rather than the score itself,
and StorePageSection's weighted sum was already float. Because the app pins
Levenshtein, fuzzywuzzy was already using the same InDel-based ratio formula
rapidfuzz implements, so scores match modulo rounding.

requirements.txt gains an explicit `rapidfuzz==3.14.3` pin (the version
already vendored transitively via Levenshtein, so the flatpak wheel set does
not churn). fuzzywuzzy and Levenshtein stay, with their comments rewritten as
ecosystem pins: the VolumeMixer plugin still imports fuzzywuzzy and ships no
requirements.txt of its own.

tests/scenario_fuzzy_ranking.py pins the contract the call sites rely on --
float range, exact/unrelated scores against every threshold in use, ranking
order, the int-returning sort comparator, and the cached staticmethod shape.

Refs #170, #166

+++ src/windows/PageManager/elements/PageSelector.py
@@ -30,7 +30,7 @@ from src.backend.DeckManagement.HelperMethods import natural_keys
 
 # Import python modules
 import os
-from fuzzywuzzy import fuzz
+from rapidfuzz import fuzz
 import re
 
 # Import globals
@@ -162,7 +162,7 @@ class PageSelector(Adw.NavigationPage):
     
     @staticmethod
     @lru_cache(maxsize=1000)
-    def calc_ratio(str1, str2) -> int:
+    def calc_ratio(str1, str2) -> float:
         # staticmethod so the cache keys only on the strings, not on self.
         return fuzz.ratio(str1.lower(), str2.lower())
